[ReFactor] player 피격 함수 playerManager 연결 / [Feat] player 아이템 줍기 수정 중

+++ Assets/Scripts/Player/PlayerCondition.cs
@@ -32,7 +32,7 @@ public class PlayerCondition : MonoBehaviour, IDamagable
         Debug.Log("Die");
     }
 
-    public void TakePhysicalDamage(int Damage)
+    public void TakePhysicalDamage(float Damage)
     {
         health.Subtract(Damage);
         onTakeDamage?.Invoke();
